Num Input Change

+++ social_app/models.py
@@ -6,7 +6,7 @@ class Profile(models.Model):
     User = models.ForeignKey(User, on_delete=models.CASCADE)
     name = models.CharField(max_length=30)
     age = models.DecimalField(decimal_places=0, max_digits=2)
-    height = models.DecimalField(decimal_places=0, max_digits=1)
+    height = models.FloatField()
     time = models.DecimalField(decimal_places=0, max_digits=4)
     email = models.CharField(max_length=100)
     def __str__(self):
